+++ app/components/Story/EndStoryControl.tsx
