+++ app/components/Story/EndStoryLayout.tsx
