+++ app/components/Story/StoryEnd.tsx
