+++ app/components/Story/StoryQuestionBox.tsx
